requirements.txt
awscli-local==0.20
boto3==1.26.30
psycopg2-binary
localstack==1.3.0
rsa